% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(data, rename_cols = NULL)
}
\arguments{
\item{data}{A data frame to be cleaned.}

\item{rename_cols}{A named vector where names are old column names and values are new names.}
}
\value{
A cleaned data frame.
}
\description{
This function cleans a dataset by removing rows with missing values and renaming columns.
}
