% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot Data}
\usage{
plot_data(data, x, y)
}
\arguments{
\item{data}{A data frame containing the variables to plot.}

\item{x}{The variable for the x-axis.}

\item{y}{The variable for the y-axis.}
}
\value{
A ggplot object representing the scatter plot with a regression line.
}
\description{
This function creates a scatter plot with a regression line.
}
